+++ .vscode/settings.json
@@ -3,7 +3,8 @@
   "editor.formatOnSave": true,
   "editor.codeActionsOnSave": {
     "source.fixAll.biome": "explicit",
-    "source.organizeImports.biome": "explicit"
+    "source.organizeImports.biome": "explicit",
+    "source.addMissingImports": "explicit"
   },
   "cSpell.words": ["sonner"]
 }
