+++ src/types/index.ts
@@ -1,29 +1,28 @@
-// Shadow settings type definitions
-export interface ShadowSettings {
+export type ShadowSettings = {
   strokeWidth: number;
   color: string;
   innerShadow: boolean;
   glow: boolean;
   glowIntensity: number;
-}
+};
 
-export interface TextSettings {
+export type TextSettings = {
   content: string;
   fontSize: number;
   fontWeight: number;
   color: string;
-}
+};
 
-export interface Preset {
+export type Preset = {
   id: string;
   name: string;
   shadowSettings: ShadowSettings;
   textSettings: TextSettings;
-}
+};
 
-export interface AppState {
+export type AppState = {
   shadowSettings: ShadowSettings;
   textSettings: TextSettings;
   presets: Preset[];
   currentPresetId: string | null;
-}
+};
