+++ src/utils/generate-id.ts
@@ -1,0 +1,3 @@
+export function generateId(): string {
+  return Date.now().toString(36) + Math.random().toString(36).substring(2);
+}
